requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.15.0
hydra-core>=1.3.2
omegaconf>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pathlib>=1.0.1
typing-extensions>=4.5.0
pyyaml>=6.0.0
logging>=0.5.0
jupyter>=1.0.0
ipython>=8.0.0
notebook>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0